realisation de part 1

+++ app/Http/Controllers/Admin/UserController.php
@@ -30,8 +30,6 @@ class UserController extends Controller
 
     public function store(Request $request){
 
-        //dd( $request->file("image")->getClientOriginalName() );
-
         $this->validate($request,[
             'name'           => 'required|string|max:255',
             'email'          => 'required|string|email|max:255|unique:users',
@@ -39,9 +37,10 @@ class UserController extends Controller
             'image'          => 'image|mimes:jpg,png,jpeg,svg|max:512|dimensions:min_width=100,min_height=100,max_width=512,max_height=512',
             'phone'          => 'required',
             'city_id'        => 'required',
-            'job_id'         => 'required'
+            'job_id'         => 'required',
+            'services'       => 'required',
+            'services.*'     => 'required',
         ]);
-
 
         $user                     = new User();
         $user->name               = $request->name;
@@ -60,8 +59,17 @@ class UserController extends Controller
             $request->file("image")->move($this->upload_path,time().$request->file("image")->getClientOriginalName());
             $user->image = time() . $request->file("image")->getClientOriginalName();
         }
+
         $user->save();
 
+        /* Les serives */
+        foreach ($request->services as $key => $value) {
+            if (isset($request->services[$key])) {
+                $user->services()->create(['name' => $request->services[$key], 'slug' => Str::slug( $request->services[$key] )]);
+            }
+        }
+
+        /* Les realisations */
         if($request->hasfile('realisations'))
         {
             foreach($request->file('realisations') as $key => $file)
@@ -91,23 +99,15 @@ class UserController extends Controller
             'password'       => 'required|string|min:8|confirmed',
             'phone'          => 'required',
             'city_id'        => 'required',
-            'job_id'         => 'required'
+            'job_id'         => 'required',
+            'services'       => 'required',
+            'services.*'     => 'required',
 
         ]);
 
         if ($request->hasFile("image")){
             $request->file("image")->move($this->upload_path,time().$request->file("image")->getClientOriginalName());
             $user->image              = time().$request->file("image")->getClientOriginalName();
-        }
-        if($request->hasfile('realisations'))
-        {
-            foreach($request->file('realisations') as $key => $file)
-            {
-                $name = time().$file->getClientOriginalName();
-                $file->move($this->realisation_path,$name);
-                $user->realisations()->create( ['name' => $name ]);
-            }
-
         }
 
         $user->name               = $request->name;
@@ -123,7 +123,29 @@ class UserController extends Controller
         $user->job_id             = $request->job_id;
         $user->save();
 
-        return redirect()->to(route('users.index'))->with('success','Modification effectué avec succès');
+        /* Les serives */
+        foreach ($request->services as $key => $value) {
+            if (isset($request->services[$key])) {
+                $user->services()->updateOrCreate(
+                    ['id'   => $request->service_id[$key] ?? null],
+                    ['name' => $request->services[$key], 'slug' => Str::slug($request->services[$key]) ]
+                );
+            }
+        }
+
+        /* Les realisations */
+        if($request->hasfile('realisations'))
+        {
+            foreach($request->file('realisations') as $key => $file)
+            {
+                $name = time().$file->getClientOriginalName();
+                $file->move($this->realisation_path,$name);
+                $user->realisations()->create( ['name' => $name ]);
+            }
+
+        }
+
+        return redirect()->back()->with('success','Modification effectué avec succès');
 
     }
 
@@ -133,12 +155,14 @@ class UserController extends Controller
 
        // if (file_exists(public_path($this->upload_path."/$user->image"))){ unlink(public_path($this->upload_path."/$user->image"));}
 
-        try{
-            if (file_exists(public_path($this->upload_path."/$user->image"))){
-                unlink(public_path($this->upload_path."/$user->image"));
+        if ($user->image != "") {
+            try {
+                if (file_exists(public_path($this->upload_path . "/$user->image"))) {
+                    unlink(public_path($this->upload_path . "/$user->image"));
+                }
+            } catch (\Exception $exception) {
+                return redirect()->back()->with('error', 'Error sur la supprission du image');
             }
-        }catch (\Exception $exception){
-            return redirect()->back()->with('error','Error sur la supprission du image');
         }
